replaced legacy hansenlaw.py with the new hansen.py version. Added docstring. Docs to be updated further
