+++ abel/__init__.py
@@ -8,7 +8,6 @@ from ._version import __version__
 from . import basex
 from . import benchmark
 from . import direct
-from . import hansen
 from . import hansenlaw
 from . import linbasex
 from . import onion_bordas
